requirements.txt
PyGithub==1.54.1
semver==2.13.0